requirements.txt
langchain==0.3.18
langchain-community==0.3.18
langchain-openai==0.3.5
langchain-text-splitters==0.3.5
chromadb==0.6.6
python-dotenv==1.0.0
tiktoken==0.12.0